Refactor models

+++ app/src/main/java/com/gabrielmorenoibarra/weatherlocation/data/api/manager/routes/WeatherApiManager.kt
@@ -11,10 +11,10 @@ class WeatherApiManager : ApiManager() {
 
     private val apiService = ApiClient.get().create(WeatherApiService::class.java)
 
-    fun get(north: Float,
-            south: Float,
-            east: Float,
-            west: Float,
+    fun get(north: Double,
+            south: Double,
+            east: Double,
+            west: Double,
             username: String,
             nPage: Int,
             nItems: Int,
